requirements.txt
aiosqlite==0.19.0
alembic==1.12.0
celery==5.3.4
greenlet==2.0.2
fastapi==0.103.1
fastapi-cache2==0.2.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
flake8==6.1.0
flower==2.0.1
isort==5.12.0
SQLAlchemy==2.0.20
uvicorn==0.23.2
